size of various types

+++ 0x00-hello_world/6-size.c
@@ -6,11 +6,11 @@
  */
 int main(void)
 {
-	printf("Size of a char:%d byte(s)\n", sizeof(char));
-	ptintf("Size of an int: %d byte(s)\n", sizeof(int));
-	printf(" Size of a long int: %d byte(s)\n", sizeof(long int));
-	printf(" Size of a long long  int: %d byte(s)\n", sizeof(long long int));
-	printf(" Size of a float: %d byte(s)\n", sizeof(float));
+	printf("Size of a char:%d byte(s)\n",(int)sizeof(char));
+	ptintf("Size of an int: %d byte(s)\n",(int)sizeof(int));
+	printf(" Size of a long int: %d byte(s)\n",(int)sizeof(long int));
+	printf(" Size of a long long  int: %d byte(s)\n",(int) sizeof(long long int));
+	printf(" Size of a float: %d byte(s)\n",(int)sizeof(float));
 	return (0);
 }
